64bit osmid support (#2422)
* first pass at support 64bit way ids again

* changelog

* back to uint64_t

* more truncated ids..

* start a test trying to validate that 64bit way ids come through. needs work in gurka though

* sort the custom way ids

* oops forgot to force extra precision flag to false until we know its true

* need double long precision on 32bits for big ids

* may as well line up all the words and save some padding

* simplify edgeinfo some more and worry about misaligned reading of extra wayid

* oops accidentally nuked text info

* fix graphtilebuilder test

* fix all the tests but restrictions

* small cleanup

* back 2 more bytes of the way id into the spare so we only need 2 extra bytes for a 64bit wayid

* lint

* Update edgeinfo.cc

* Update edgeinfo.cc

* Update edgeinfo.h

* try robin hood for uncapped nodeids

* allow nodes ids up to 64bits. strengthing idtable test

* add getting to the test as well

* also check for ids that arent there

* changelog

* updated to kMaxOSMNodesHint

* missed a few.  32->64

* should not of updated hasher with last update.

* kick ci

* TODOs describing how to completely remove the bitsets

* do away with all usage of idtable in tile building

* dont need this constant and straighten up the edge counting

* redundant

* simplify intersection matching test because its relying on fiddly floating point equality

* fix removed print of nodes processed

* we shouldnt blacklist the first edge info in the tile

* add complex restriction test for restrictions regression

* more uint64_t updates.

* remove print

* using higher ids causes restriction builder to miss some restrictions!

* mised another blacklisted offset

* test more parts of the wayid space

* add a warning about which way ids were marked as a restriction but werent consequtive

* match the actual callback signature

* failing complex restrictions with randomly generated way ids

* simplify the ids to just the single via that cannot be found

* add to 64bit way id test to prove that the way id comes out wrong from the edgeinfo

* fix edgeinfo and also use 1 less byte where possible. restrictions are fixed now and were just a symptom of edgeinfo wayid bugs

* disable restriction walking warning

* Update graphbuilder.cc

* wrong namespace

+++ src/mjolnir/graphfilter.cc
@@ -80,7 +80,7 @@ void FilterTiles(GraphReader& reader,
       uint32_t edge_index = tilebuilder.directededges().size();
 
       // Iterate through directed edges outbound from this node
-      std::vector<uint32_t> wayid;
+      std::vector<uint64_t> wayid;
       std::vector<GraphId> endnode;
       const NodeInfo* nodeinfo = tile->node(nodeid);
       GraphId edgeid(nodeid.tileid(), nodeid.level(), nodeinfo->edge_index());
